Make ip address/port optional
Some participants of l2mesh networks are not publicly reachable from the
internet. Those hosts should not lead to a ConnectTO entry in the other
hosts' configurations.
This change adds support for this scenario by simply not specifying an
IP address for those hosts.

Additionally, it makes the port setting optional, which causes tinc to
default to the standard 655 port.